+++ contracts/interfaces/IBridge.sol
@@ -9,5 +9,5 @@ interface IBridge {
         @notice Exposing getter for {_chainID} instead of forcing the use of call.
         @return uint8 The {_chainID} that is currently set for the Bridge contract.
      */
-    function _chainID() external returns (uint8);
+    function _chainID() external returns (uint32);
 }
